Add files via upload
Challenge Description: This POAP design will be used for the Weekly Forum AMA's for a month. This POAP will be delivered to all those who attend the Forum call in Discord or Twitter Space (Twitter pending on deliver mechanisms). This design needs to offer the ability to have the Episode number editable for each Episode.